requirements.txt
alpaca-py>=0.13.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pytz>=2024.1
